build(Gradle): Migrate application declarations to conventions
