+++ cli/build.gradle.kts
@@ -17,112 +17,14 @@
  * License-Filename: LICENSE
  */
 
-import java.nio.charset.Charset
-import java.nio.file.Files
-
-import org.graalvm.buildtools.gradle.tasks.BuildNativeImageTask
-
 plugins {
     // Apply precompiled plugins.
     id("ort-application-conventions")
-
-    // Apply third-party plugins.
-    alias(libs.plugins.graalVmNativeImage)
 }
 
 application {
     applicationName = "ort"
-    applicationDefaultJvmArgs = listOf(
-        "--add-opens", "java.base/sun.nio.ch=ALL-UNNAMED",
-        "--add-opens", "java.base/java.io=ALL-UNNAMED"
-    )
     mainClass.set("org.ossreviewtoolkit.cli.OrtMainKt")
-}
-
-graalvmNative {
-    toolchainDetection.set(true)
-
-    // For options see https://graalvm.github.io/native-build-tools/latest/gradle-plugin.html.
-    binaries {
-        named("main") {
-            imageName.set("ort")
-
-            val initializeAtBuildTime = listOf(
-                "ch.qos.logback.classic.Level",
-                "ch.qos.logback.classic.Logger",
-                "ch.qos.logback.classic.PatternLayout",
-                "ch.qos.logback.core.CoreConstants",
-                "ch.qos.logback.core.status.InfoStatus",
-                "ch.qos.logback.core.status.StatusBase",
-                "ch.qos.logback.core.util.Loader",
-                "ch.qos.logback.core.util.StatusPrinter",
-                "org.apache.sshd.sftp.client.fs.SftpFileSystemProvider",
-                "org.slf4j.LoggerFactory"
-            ).joinToString(separator = ",", prefix = "--initialize-at-build-time=")
-
-            buildArgs.add(initializeAtBuildTime)
-        }
-    }
-
-    metadataRepository {
-        enabled.set(true)
-    }
-}
-
-// JDK 20 will only be supported starting with GraalVM 23, see
-// https://www.graalvm.org/release-notes/release-calendar/#planned-releases
-if (JavaVersion.current().majorVersion.toInt() <= 19) {
-    tasks.named<BuildNativeImageTask>("nativeCompile") {
-        // Gradle's "Copy" task cannot handle symbolic links, see https://github.com/gradle/gradle/issues/3982. That is
-        // why links contained in the GraalVM distribution archive get broken during provisioning and are replaced by
-        // empty files. Address this by recreating the links in the toolchain directory.
-        val toolchainDir = options.get().javaLauncher.get().executablePath.asFile.parentFile.run {
-            if (name == "bin") parentFile else this
-        }
-
-        val toolchainFiles = toolchainDir.walkTopDown().filter { it.isFile }
-        val emptyFiles = toolchainFiles.filter { it.length() == 0L }
-
-        // Find empty toolchain files that are named like other toolchain files and assume these should have been links.
-        val links = toolchainFiles.mapNotNull { file ->
-            emptyFiles.singleOrNull { it != file && it.name == file.name }?.let {
-                file to it
-            }
-        }
-
-        // Fix up symbolic links.
-        links.forEach { (target, link) ->
-            logger.quiet("Fixing up '$link' to link to '$target'.")
-
-            if (link.delete()) {
-                Files.createSymbolicLink(link.toPath(), target.toPath())
-            } else {
-                logger.warn("Unable to delete '$link'.")
-            }
-        }
-    }
-}
-
-tasks.named<CreateStartScripts>("startScripts") {
-    doLast {
-        // Work around the command line length limit on Windows when passing the classpath to Java, see
-        // https://github.com/gradle/gradle/issues/1989#issuecomment-395001392.
-        val windowsScriptText = windowsScript.readText(Charset.defaultCharset())
-        windowsScript.writeText(
-            windowsScriptText.replace(
-                Regex("set CLASSPATH=%APP_HOME%\\\\lib\\\\.*"),
-                "set CLASSPATH=%APP_HOME%\\\\lib\\\\*;%APP_HOME%\\\\plugin\\\\*"
-            )
-        )
-
-        val unixScriptText = unixScript.readText(Charset.defaultCharset())
-        unixScript.writeText(
-            unixScriptText.replace(
-                Regex("CLASSPATH=\\\$APP_HOME/lib/.*"),
-                "CLASSPATH=\\\$APP_HOME/lib/*:\\\$APP_HOME/plugin/*"
-            )
-        )
-    }
 }
 
 dependencies {
