+++ helper-cli/build.gradle.kts
@@ -17,8 +17,6 @@
  * License-Filename: LICENSE
  */
 
-import java.nio.charset.Charset
-
 plugins {
     // Apply precompiled plugins.
     id("ort-application-conventions")
@@ -26,33 +24,7 @@ plugins {
 
 application {
     applicationName = "orth"
-    applicationDefaultJvmArgs = listOf(
-        "--add-opens", "java.base/sun.nio.ch=ALL-UNNAMED",
-        "--add-opens", "java.base/java.io=ALL-UNNAMED"
-    )
     mainClass.set("org.ossreviewtoolkit.helper.HelperMainKt")
-}
-
-tasks.named<CreateStartScripts>("startScripts") {
-    doLast {
-        // Work around the command line length limit on Windows when passing the classpath to Java, see
-        // https://github.com/gradle/gradle/issues/1989#issuecomment-395001392.
-        val windowsScriptText = windowsScript.readText(Charset.defaultCharset())
-        windowsScript.writeText(
-            windowsScriptText.replace(
-                Regex("set CLASSPATH=%APP_HOME%\\\\lib\\\\.*"),
-                "set CLASSPATH=%APP_HOME%\\\\lib\\\\*;%APP_HOME%\\\\plugin\\\\*"
-            )
-        )
-
-        val unixScriptText = unixScript.readText(Charset.defaultCharset())
-        unixScript.writeText(
-            unixScriptText.replace(
-                Regex("CLASSPATH=\\\$APP_HOME/lib/.*"),
-                "CLASSPATH=\\\$APP_HOME/lib/*:\\\$APP_HOME/plugin/*"
-            )
-        )
-    }
 }
 
 dependencies {
